chore: separate out config types into modules
