+++ src/modules/candy-machine/models/config/Creators.ts
@@ -1,0 +1,7 @@
+import { Creator } from '@metaplex-foundation/mpl-token-metadata';
+import { PublicKeyString } from '../../../../shared';
+
+export type CreatorConfig = Creator & {
+  address: PublicKeyString;
+};
+export type CreatorsConfig = CreatorConfig[];
