+++ src/modules/candy-machine/models/config/Gatekeeper.ts
@@ -1,0 +1,10 @@
+/**
+ * Configures {@link CandyMachineConfig.gatekeeper} settings.
+ *
+ * @property gatekeeperNetwork - Gateway provider address
+ * @property expireOnUse - Requires a new gateway challenge after a use
+ */
+export type GatekeeperSettingsConfig = {
+  gatekeeperNetwork: string;
+  expireOnUse: boolean;
+};
